tdd-requirements.txt
setuptools
wheel
twine
pytest
pillow
opencv-python
numpy